rename body package to connutil
