+++ internal/pkg/archiver/connutil/connutil.go
@@ -1,4 +1,4 @@
-package body
+package connutil
 
 import (
 	"errors"
